backend/orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
google-genai==0.3.0
google-cloud-firestore==2.20.1
google-cloud-storage==2.19.0
google-cloud-tasks==2.17.0
firebase-admin==6.6.0
redis==5.2.1
python-dotenv==1.0.1
pytest==8.3.4

